feat(storybook-templates): created storybook templates and updated old rhc-templates to get them from storybook
