+++ apps/rhc-templates/package.json
@@ -37,6 +37,7 @@
     "@rijkshuisstijl-community/design-tokens": "workspace:*",
     "@rijkshuisstijl-community/font": "workspace:*",
     "@rijkshuisstijl-community/rivm-design-tokens": "workspace:*",
+    "@rijkshuisstijl-community/storybook": "workspace:*",
     "@rijkshuisstijl-community/web-components-react": "workspace:*",
     "@rijkshuisstijl-community/web-components-stencil": "workspace:*",
     "@tabler/icons-react": "3.11.0",
