+++ packages/components-react/src/index.ts
@@ -26,6 +26,8 @@ export {
 } from './BreadcrumbNav';
 export {
   Button,
+  ButtonLink,
+  type ButtonLinkProps,
   type ButtonProps,
   IconButton,
   PrimaryActionButton,
