+++ packages/storybook/package.json
@@ -36,7 +36,6 @@
     "@rijkshuisstijl-community/font": "workspace:*",
     "@rijkshuisstijl-community/logius-design-tokens": "workspace:*",
     "@rijkshuisstijl-community/mijnoverheid-design-tokens": "workspace:*",
-    "@rijkshuisstijl-community/rhc-templates": "workspace:*",
     "@rijkshuisstijl-community/rivm-design-tokens": "workspace:*",
     "@rijkshuisstijl-community/web-components-react": "workspace:*",
     "@rijkshuisstijl-community/web-components-stencil": "workspace:*",
