+++ packages/storybook/src/templates/details.stories.tsx
@@ -1,13 +1,14 @@
 import type { Meta, StoryObj } from '@storybook/react/*';
-import Home from './home';
+import Page from './details';
+import './globals.css';
 
 const meta: Meta = {
-  title: 'Templates/Home',
-  component: Home,
+  title: 'Templates/Details',
+  component: Page,
   parameters: {},
 };
 export default meta;
 
-type Story = StoryObj<typeof Home>;
+type Story = StoryObj<typeof Page>;
 
 export const Default: Story = {};
