+++ packages/storybook/src/templates/globals.css
@@ -1,76 +1,65 @@
 /* global CSS here */
 body {
-  margin: 0;
   display: flex;
   flex-direction: column;
-  padding: 0;
+  margin-block: 0;
+  margin-inline: 0;
+  padding-block: 0;
+  padding-inline: 0;
 }
 
-.page {
+.rhc-templates-page {
   background-color: #f2f4f6;
 }
 
-.unstarted {
+.rhc-templates-unstarted {
   border: 2px dashed red;
 }
 
-.unfinished {
+.rhc-templates-unfinished {
   border: 2px dashed yellow;
 }
 
-.explanation {
+.rhc-templates-explanation {
   color: red;
 }
-.dutch-map {
+.rhc-templates-dutch-map {
   --utrecht-icon-size: 48px;
 }
 
-.rhc-background-color-white {
+.rhc-templates-background-color-white {
   background-color: var(--rhc-color-wit);
 }
 
-.rhc-background-color-lint-blauw {
+.rhc-templates-background-color-lint-blauw {
   background-color: var(--rhc-color-lintblauw-500);
 }
 
-.rhc-spacing-2 {
+.rhc-templates-spacing-2 {
   margin-block-end: var(--rhc-space-400);
 }
 
-.rhc-main-content {
+.rhc-templates-main-content {
+  align-items: center;
   display: flex;
   justify-content: center;
-  align-items: center;
 }
 
-.utrecht-figure {
-  margin-inline-start: 0;
-  margin-inline-end: 0;
-}
-
-.rhc-page-content {
+.rhc-templates-page-content {
   --utrecht-page-content-padding-block-end: var(--rhc-space-600);
   --utrecht-page-content-padding-block-start: var(--rhc-space-600);
   --utrecht-page-padding-inline-end: var(--rhc-space-700);
   --utrecht-page-padding-inline-start: var(--rhc-space-700);
 
-  margin-block-start: 80px;
+  inline-size: 70%;
   margin-block-end: 80px;
-  width: 70%;
+  margin-block-start: 80px;
 }
 
-.utrecht-nav-list {
-  gap: 0.5rem;
-}
-
-.utrecht-nav-list__link a {
-  color: white;
-}
-
-@media (max-width: 780px) {
-  .rhc-page-content {
-    width: 100%;
-    margin-block-start: 0;
+@media (width <= 780px) {
+  .rhc-templates-page-content {
+    inline-size: 100%;
     margin-block-end: 0;
+    margin-block-start: 0;
   }
 }
