--- packages/storybook/src/templates/home.tsx
@@ -1,12 +1,0 @@
-import RootLayout from '@rijkshuisstijl-community/rhc-templates/src/app/layout';
-import Page from '@rijkshuisstijl-community/rhc-templates/src/app/page';
-
-export default function Home() {
-  return (
-    <>
-      <RootLayout>
-        <Page />
-      </RootLayout>
-    </>
-  );
-}
