--- packages/storybook/src/templates/layouts/layout.tsx
@@ -1,40 +1,0 @@
-'use client';
-import { Document } from '@utrecht/component-library-react';
-import { PropsWithChildren, ReactNode } from 'react';
-import '@nl-rvo/assets/fonts/index.css';
-import '@rijkshuisstijl-community/design-tokens/dist/index.css';
-import '@rijkshuisstijl-community/components-css/index.scss';
-import { Logo, Icon, NavBar, PageHeader } from '@rijkshuisstijl-community/components-react';
-import '@rijkshuisstijl-community/font/src/index.mjs';
-import './globals.css';
-
-const RHCTheme = ({ children }: { children: ReactNode }) => <div className="rhc-theme">{children}</div>;
-
-export default function RootLayout({ children }: PropsWithChildren<{}>) {
-  return (
-    <html lang="nl">
-      <body>
-        <RHCTheme>
-          <div className="rhc-background-color-white">
-            <PageHeader>
-              <Logo organisation="Rijkshuisstijl Community">
-                <Icon icon={'nederland-map'} className={'dutch-map'} />
-              </Logo>
-            </PageHeader>
-          </div>
-          <NavBar
-            items={[
-              { href: '/', label: 'Page' },
-              { href: '/form', label: 'Form' },
-              { href: '/details', label: 'Details' },
-              { href: '/collage', label: 'Collage' },
-            ]}
-          ></NavBar>
-          <main className="page">
-            <Document>{children}</Document>
-          </main>
-        </RHCTheme>
-      </body>
-    </html>
-  );
-}
